fix: Double type for started_at (backend returns float timestamps)

+++ app/src/main/java/com/m57/hermescontrol/data/model/ActivityItem.kt
@@ -11,5 +11,5 @@ data class ActivityItem(
     val status: String,
     val formattedTime: String,
     val messageCount: Int,
-    val timestamp: Long,
+    val timestamp: Double,
 )
